upload-backend/monprojet/requirements.txt
anytree==2.12.1
asgiref==3.11.1
certifi==2026.1.4
customtkinter==5.2.2
darkdetect==0.8.0
distlib==0.4.0
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.3
packaging==24.2
pipenv==2026.0.3
platformdirs==4.5.1
PyJWT==2.11.0
PyMySQL==1.1.2
python-decouple==3.8
setuptools==82.0.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
virtualenv==20.36.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0